filterable charts
this is what it's all about!!! very happy I finally got these, it really
is as interesting to filter as I imagined. def could use a bit of work
wrt UX but on the whole this is exactly the pattern I had imagined.

Now I just need to make it so the parquet files are generated
automatically...