use interaction.followup to avoid timeouts
turns out, discord wants interaction.respond to be called fairly
quickly. With my more elaborate blog setup (where it was building
locally instead of waiting for GitHub), this was not happening, so I've
modified this code to send a message earlier and edit it later.

thank u discord.py for being a very easy-to-use and feature-complete
library